Reorganized some classes into new model.declaration package.
To make it more clear these classes are tightly related.
