+++ src/main/kotlin/com/autonomousapps/model/DependencyGraphView.kt
@@ -4,7 +4,7 @@
 package com.autonomousapps.model
 
 import com.autonomousapps.internal.unsafeLazy
-import com.autonomousapps.model.intermediates.Variant
+import com.autonomousapps.model.declaration.Variant
 import com.google.common.graph.ElementOrder
 import com.google.common.graph.Graph
 import com.google.common.graph.GraphBuilder
