+++ src/main/kotlin/com/autonomousapps/visitor/GraphViewVisitor.kt
@@ -3,7 +3,7 @@ package com.autonomousapps.visitor
 import com.autonomousapps.model.Dependency
 import com.autonomousapps.model.DependencyGraphView
 import com.autonomousapps.model.ProjectVariant
-import com.autonomousapps.model.intermediates.Declaration
+import com.autonomousapps.model.declaration.Declaration
 
 internal interface GraphViewVisitor {
   fun visit(dependency: Dependency, context: Context)
